+++ changelog.md
@@ -2,6 +2,18 @@
 
 stack-analyze desktop all version and notable changes, fixed, remove and new additions in code.
 
+## version 3.0.0
+### Added
+- add new tools
+    - anime search
+    - github info
+- add recomendation section from version 1.0.8 (1.0.85) - 1.0.9
+### Fixed
+- redisgn html validator with hybrid mode (css native and materialize)
+### Changed
+- pagespeed is separate the tech-stack tool now pagespeed is a tool standalone
+- change the doughnut chart to bar chart
+
 ## version 2.0.0
 ### Added
 - add the new modules:
@@ -17,7 +29,7 @@ stack-analyze desktop all version and notable changes, fixed, remove and new add
 
 ## version 1.0.0
 - fisrt official version in npm with the modules:
-- main modules:    
+- main modules:
     - materialize-css
     - material-design-icons
     - wappalyzer